rename edit action
