+++ src/ducks/Watchlists/Actions/Edit/AssetsList.module.scss
